lambda/ocr_processor/requirements.txt
boto3>=1.34.0
pytesseract>=0.3.10
Pillow>=10.0.0
